Change schedule to 500Hz
Note: In the previous commit there was an erroneous mention of a sensor
inconsistency. This was caused by applying motion to only 1 of the 3
sets of IMUs causing them to not align and throw an error. There does
not appear to be any issue when they are all mounted on the same frame.